move PlotlySurfaceFigfure test
