+++ nilearn/plotting/surface/tests/test_plotly_backend.py
@@ -219,150 +219,20 @@ def test_configure_title():
     assert config["font"]["color"] == "green"
 
 
-def test_value_error_get_faces_on_edge(plotly, in_memory_mesh):
-    """Test that calling _get_faces_on_edge raises a ValueError when \
-       called with with indices that do not form a region.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError, match=("Vertices in parcellation do not form region.")
-    ):
-        figure._get_faces_on_edge([91])
-
-
-def test_plot_surf_contours_errors_with_plotly_figure(plotly, in_memory_mesh):
+def test_plot_surf_contours_errors_with_plotly_figure(in_memory_mesh):
     """Test that plot_surf_contours rasises error when given plotly obj."""
-    figure = plot_surf(in_memory_mesh, engine="plotly")
+    figure = plot_surf(in_memory_mesh, engine=ENGINE)
     with pytest.raises(ValueError):
         plot_surf_contours(in_memory_mesh, np.ones((10,)), figure=figure)
 
 
-def test_plot_surf_contours_errors_with_plotly_axes(plotly, in_memory_mesh):
+def test_plot_surf_contours_errors_with_plotly_axes(in_memory_mesh):
     """Test that plot_surf_contours rasises error when given plotly \
         obj as axis.
     """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
+    figure = plot_surf(in_memory_mesh, engine=ENGINE)
     with pytest.raises(ValueError):
         plot_surf_contours(in_memory_mesh, np.ones((10,)), axes=figure)
-
-
-def test_plotly_surface_figure_warns_on_isolated_roi(plotly, in_memory_mesh):
-    """Test that a warning is generated for ROIs with isolated vertices."""
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    # the method raises an error because the (randomly generated)
-    # vertices don't form regions
-    try:
-        with pytest.raises(UserWarning, match="contains isolated vertices:"):
-            figure.add_contours(levels=[0], roi_map=np.array([0, 1] * 10))
-    except Exception:
-        pass
-
-
-@pytest.mark.parametrize("levels,labels", [([0], ["a", "b"]), ([0, 1], ["a"])])
-def test_value_error_add_contours_levels_labels(
-    plotly, levels, labels, in_memory_mesh
-):
-    """Test that add_contours raises a ValueError when called with levels and \
-    labels that have incompatible lengths.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError,
-        match=("levels and labels need to be either the same length or None."),
-    ):
-        figure.add_contours(
-            levels=levels, labels=labels, roi_map=np.ones((10,))
-        )
-
-
-@pytest.mark.parametrize(
-    "levels,lines",
-    [([0], [{}, {}]), ([0, 1], [{}, {}, {}])],
-)
-def test_value_error_add_contours_levels_lines(
-    plotly, levels, lines, in_memory_mesh
-):
-    """Test that add_contours raises a ValueError when called with levels and \
-    lines that have incompatible lengths.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError,
-        match=("levels and lines need to be either the same length or None."),
-    ):
-        figure.add_contours(levels=levels, lines=lines, roi_map=np.ones((10,)))
-
-
-def test_add_contours(plotly, surface_image_roi):
-    """Test that add_contours updates data in PlotlySurfaceFigure."""
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi)
-    assert len(figure.figure.to_dict().get("data")) == 4
-
-    figure.add_contours(surface_image_roi, levels=[1])
-    assert len(figure.figure.to_dict().get("data")) == 5
-
-
-@pytest.mark.parametrize("hemi", ["left", "right", "both"])
-def test_add_contours_hemi(
-    plotly,
-    surface_image_roi,
-    hemi,
-):
-    """Test that add_contours works with all hemi inputs."""
-    if hemi == "both":
-        n_vertices = surface_image_roi.mesh.n_vertices
-    else:
-        n_vertices = surface_image_roi.data.parts[hemi].shape[0]
-    figure = plot_surf(
-        surface_image_roi.mesh,
-        engine="plotly",
-        hemi=hemi,
-    )
-    figure.add_contours(surface_image_roi)
-    assert figure._coords.shape[0] == n_vertices
-
-
-def test_add_contours_plotly_surface_image(plotly, surface_image_roi):
-    """Test that add_contours works with SurfaceImage."""
-    figure = plot_surf(
-        surf_map=surface_image_roi, hemi="left", engine="plotly"
-    )
-    figure.add_contours(roi_map=surface_image_roi)
-
-
-def test_add_contours_has_name(plotly, surface_image_roi):
-    """Test that contours added to a PlotlySurfaceFigure can be named."""
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, levels=[1], labels=["x"])
-    assert figure.figure.to_dict().get("data")[2].get("name") == "x"
-
-
-def test_add_contours_lines_duplicated(plotly, surface_image_roi):
-    """Test that the specifications of length 1 line provided to \
-     add_contours are duplicated to all requested contours.
-    """
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, lines=[{"width": 10}])
-    newlines = figure.figure.to_dict().get("data")[2:]
-    assert all(x.get("line").__contains__("width") for x in newlines)
-
-
-@pytest.mark.parametrize(
-    "key,value",
-    [
-        ("color", "yellow"),
-        ("width", 10),
-    ],
-)
-def test_add_contours_line_properties(plotly, key, value, surface_image_roi):
-    """Test that the specifications of a line provided to add_contours are \
-    stored in the PlotlySurfaceFigure data.
-    """
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, levels=[1], lines=[{key: value}])
-    newline = figure.figure.to_dict().get("data")[2].get("line")
-    assert newline.get(key) == value
 
 
 @pytest.mark.skipif(
